Moved leftover SuiteSparse source classes to win64 project
